requirements.txt
python-dotenv
openai
langchain
pypdf
faiss-cpu
tiktoken